Routing index graph test refactoring.
