+++ routing/routing_tests/index_graph_tools.hpp
@@ -1,0 +1,49 @@
+#pragma once
+
+#include "routing/edge_estimator.hpp"
+#include "routing/index_graph.hpp"
+#include "routing/index_graph_starter.hpp"
+#include "routing/road_point.hpp"
+
+#include "routing/base/astar_algorithm.hpp"
+
+#include "geometry/point2d.hpp"
+
+#include "std/algorithm.hpp"
+#include "std/shared_ptr.hpp"
+#include "std/unique_ptr.hpp"
+#include "std/unordered_map.hpp"
+#include "std/vector.hpp"
+
+namespace routing_test
+{
+class TestGeometryLoader final : public routing::GeometryLoader
+{
+public:
+  // GeometryLoader overrides:
+  void Load(uint32_t featureId, routing::RoadGeometry & road) const override;
+
+  void AddRoad(uint32_t featureId, bool oneWay, float speed,
+               routing::RoadGeometry::Points const & points);
+
+private:
+  unordered_map<uint32_t, routing::RoadGeometry> m_roads;
+};
+
+routing::Joint MakeJoint(vector<routing::RoadPoint> const & points);
+
+shared_ptr<routing::EdgeEstimator> CreateEstimator();
+
+routing::AStarAlgorithm<routing::IndexGraphStarter>::Result CalculateRoute(
+    routing::IndexGraphStarter & graph, vector<routing::RoadPoint> & roadPoints);
+
+void TestRouteSegments(
+    routing::IndexGraphStarter & starter,
+    routing::AStarAlgorithm<routing::IndexGraphStarter>::Result expectedRouteResult,
+    vector<routing::RoadPoint> const & expectedRoute);
+
+void TestRouteGeometry(
+    routing::IndexGraphStarter & starter,
+    routing::AStarAlgorithm<routing::IndexGraphStarter>::Result expectedRouteResult,
+    vector<m2::PointD> const & expectedRouteGeom);
+}  // namespace routing_test
