+++ routing/routing_tests/routing_tests.pro
@@ -29,6 +29,7 @@ SOURCES += \
   cross_routing_tests.cpp \
   followed_polyline_test.cpp \
   index_graph_test.cpp \
+  index_graph_tools.cpp \
   nearest_edge_finder_tests.cpp \
   online_cross_fetcher_test.cpp \
   osrm_router_test.cpp \
@@ -43,4 +44,5 @@ SOURCES += \
   vehicle_model_test.cpp \
 
 HEADERS += \
+  index_graph_tools.hpp \
   road_graph_builder.hpp \
